temp/bridge_codex_copilot_bridge/services/ita/pyproject.toml

[project]
name = "ita"
version = "0.1.0"
description = "Internal Tools API for Codex & Copilot bridge"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.115.2",
  "uvicorn[standard]==0.30.6",
  "pydantic==2.9.2",
  "httpx==0.27.2",
  "PyJWT==2.9.0",
  "cryptography==43.0.1",
]

[tool.pytest.ini_options]
addopts = "-q"
